Feature: Added synchronization between the inventory and the main screen. Now, when you click on one of the three buttons on the main screen, "Feed, Water, Play", it will deduct the item related to the subsequent button pressed.
Feed = Apples
Water = Water
Play = Ball

Added snackbar error messages on insufficient items to complete task.

Added a "Reset pet stats" button for testing purposes.

Encountered a bug that would not increment the stats of the tamagotchi, only on the FIRST click of any button.

Fixed the "first-click" bug by refactoring the TamagotchiViewModel to use a single StateFlow for its UI state, which ensures a single source of truth for the UI.